Separate Adguard code
